backend/database/models.py
from pydantic import BaseModel, Field, EmailStr
from typing import Optional, List
from datetime import datetime
from bson import ObjectId

class PyObjectId(ObjectId):
    @classmethod
    def __get_validators__(cls):
        yield cls.validate

    @classmethod
    def validate(cls, v):
        if not ObjectId.is_valid(v):
            raise ValueError("Invalid objectid")
        return ObjectId(v)

    @classmethod
    def __modify_schema__(cls, field_schema):
        field_schema.update(type="string")

class MongoBaseModel(BaseModel):
    id: Optional[PyObjectId] = Field(default_factory=PyObjectId, alias="_id")

    class Config:
        allow_population_by_field_name = True
        arbitrary_types_allowed = True
        json_encoders = {ObjectId: str}

class User(MongoBaseModel):
    full_name: str
    email: EmailStr
    hashed_password: str
    role: str = "member"  # member, admin, ai_head
    face_encoding: Optional[List[float]] = None
    is_active: bool = True
    created_at: datetime = Field(default_factory=datetime.utcnow)

class Attendance(MongoBaseModel):
    user_id: str
    date: datetime
    status: str  # present, absent, late
    check_in: Optional[datetime] = None
    check_out: Optional[datetime] = None
    confidence: float = 0.0
