requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pymongo
motor
python-dotenv
face_recognition
numpy
opencv-python
pandas
openpyxl
reportlab
python-socketio
aiohttp
google-auth
google-auth-oauthlib
google-auth-httplib2
openai
google-generativeai
scikit-learn
pydantic-settings
pytz
